backend/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
ollama>=0.2.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
